rename videoCourse folder
