+++ __tests__/snapshot/videoPlayer.test.js
@@ -1,6 +1,6 @@
 import { create } from "react-test-renderer";
 import { ThemeProvider } from "styled-components";
-import VideoPlayer from "../../src/components/videoCourses/videoPlayer";
+import VideoPlayer from "../../src/components/videoCourse/videoPlayer";
 
 const theme = {
   colors: {
