+++ src/components/videoCourse/courseDetail.tsx
