+++ src/pages/myCourses/classRoom/videoCourseRoom/[courseId].tsx
@@ -6,8 +6,8 @@ import Swal from "sweetalert2";
 import Head from "next/head";
 import { getUserData, getVideoCourse } from "../../../../utils/firestore";
 import { AuthContext } from "../../../../contexts/authContext";
-import VideoPlayer from "../../../../components/videoCourses/videoPlayer";
-import CourseDetail from "../../../../components/videoCourses/courseDetail";
+import VideoPlayer from "../../../../components/videoCourse/videoPlayer";
+import CourseDetail from "../../../../components/videoCourse/courseDetail";
 import Play from "../../../../../public/play.png";
 import Loading from "../../../../../public/loading.gif";
 import "react-quill/dist/quill.snow.css";
