+++ src/pages/videoCourses/courseDetail/[courseId].tsx
@@ -8,8 +8,8 @@ import { useRecoilState } from "recoil";
 import { getVideoCourse, getUserData, updateCartItem } from "../../../utils/firestore";
 import { AuthContext } from "../../../contexts/authContext";
 import { orderQtyState, showMemberModalState } from "../../../utils/recoil";
-import VideoPlayer from "../../../components/videoCourses/videoPlayer";
-import CourseDetail from "../../../components/videoCourses/courseDetail";
+import VideoPlayer from "../../../components/videoCourse/videoPlayer";
+import CourseDetail from "../../../components/videoCourse/courseDetail";
 import Lock from "../../../../public/lock.png";
 import Play from "../../../../public/play.png";
 import "react-quill/dist/quill.snow.css";
